refactor(run): collectFees as a shared library
* more collectFees to a shared library
* reuse it in `vault` and `psm`
* add PSM test for `collectFees`
